rvv: add saturation helper function
1. add integer signed add/subu and unsigend add/sub saturation function
2. merge these with mulhi helper
