+++ riscv/insn_template.h
@@ -1,7 +1,7 @@
 // See LICENSE for license details.
 
+#include "arith.h"
 #include "mmu.h"
-#include "mulhi.h"
 #include "softfloat.h"
 #include "internals.h"
 #include "specialize.h"
