--- riscv/mulhi.h
@@ -1,43 +1,0 @@
-// See LICENSE for license details.
-
-#ifndef _RISCV_MULHI_H
-#define _RISCV_MULHI_H
-
-#include <cstdint>
-
-inline uint64_t mulhu(uint64_t a, uint64_t b)
-{
-  uint64_t t;
-  uint32_t y1, y2, y3;
-  uint64_t a0 = (uint32_t)a, a1 = a >> 32;
-  uint64_t b0 = (uint32_t)b, b1 = b >> 32;
-
-  t = a1*b0 + ((a0*b0) >> 32);
-  y1 = t;
-  y2 = t >> 32;
-
-  t = a0*b1 + y1;
-  y1 = t;
-
-  t = a1*b1 + y2 + (t >> 32);
-  y2 = t;
-  y3 = t >> 32;
-
-  return ((uint64_t)y3 << 32) | y2;
-}
-
-inline int64_t mulh(int64_t a, int64_t b)
-{
-  int negate = (a < 0) != (b < 0);
-  uint64_t res = mulhu(a < 0 ? -a : a, b < 0 ? -b : b);
-  return negate ? ~res + (a * b == 0) : res;
-}
-
-inline int64_t mulhsu(int64_t a, uint64_t b)
-{
-  int negate = a < 0;
-  uint64_t res = mulhu(a < 0 ? -a : a, b);
-  return negate ? ~res + (a * b == 0) : res;
-}
-
-#endif
